Fiixes #243: Same AppBar component for all routes
